trino: Transform ARRAY(SELECT ...) to TRANSFORM
Before, we rewrote this to a subquery, which worked fine at the top
level. But when nested, it failed with an error about an unsupported
correlated subquery.

This version heavily restricts the grammar of `ARRAY(SELECT ...)`,
limiting it to use cases that resemble an "array comprehension". Then we
transform these into some combination of `TRANSFORM`, `FILTER`, and
`ARRAY_DISTINCT`.

This involves a fair bit of work and complexity, but it allows
BigQuery's standard substitues for list comprehensions to work as
expected on Trino. Similar strategies are likely to be needed for
Snowflake.

We also fix a few issues here and there, and add a type declaration for
the `MOD` function.