Add ConnectorType to Target::Trino
We will need to distinguish between Trino connectors like `memory`,
`hive` and `iceberg` when emitting Trino SQL, because the different
connectors support subsets of Trino's features and types.

We modify the `Target` enum to include a `Target::Trino(ConnectionType)`
variant instead of just `Target`. This will allow `emit` code to know
what target it is generating code for.

This also removes the final pretense of a "standalone" mode where we
can run without access to a specific Trino server. This is fine; we're
moving in that direction anyway.

Note the replacement of `t == Target::Trino` with `matches!(t,
Target::Trino(_))`, which is a handy Rust feature.