WIP: Implement LOAD operations
This turned out to be surprisingly tricky. The key insight is that LOAD
and STORE aren't really symmetrical, at least not in this version of the
code. Instead, %STORE(expr) gets manually inserted as a phantom
operator, but %LOAD(var_ref) only occurs as part of a variable
reference.

This means that transforms, `sql_quote!` and all our other rewriting
machinery don't even notice that LOADs exist.

This is a weird decision, and it might not be optimal in the long run,
but it works.

This brings us down to `7 tests failed, 64 passed, 3 pending`, and
all the failing test cases are places we need to expand SELECT wildcards
like `*`, `table.*`, etc., into column lists with explicit STORE
operations.